Add indice in all normas and change integer type by bigint postgres

+++ db/migrate/20150810124203_add_column_indice_to_normas.rb
@@ -1,5 +1,5 @@
 class AddColumnIndiceToNormas < ActiveRecord::Migration
   def change
-    add_column :normas, :indice, :integer
+    add_column :normas, :indice, :bigint
   end
 end
